refactor(matrices): move copy, abs, has to RepMatrix
